add button function

+++ pixel_ring/record.py
@@ -1,6 +1,45 @@
 import pyaudio
 import wave
 
+import digitalio
+import board
+
+from adafruit_rgb_display.rgb import color565
+import adafruit_rgb_display.st7789 as st7789
+import webcolors
+
+
+# The display uses a communication protocol called SPI.
+# SPI will not be covered in depth in this course. 
+# you can read more https://www.circuitbasics.com/basics-of-the-spi-communication-protocol/
+cs_pin = digitalio.DigitalInOut(board.CE0)
+dc_pin = digitalio.DigitalInOut(board.D25)
+reset_pin = None
+BAUDRATE = 64000000  # the rate  the screen talks to the pi
+# Create the ST7789 display:
+display = st7789.ST7789(
+    board.SPI(),
+    cs=cs_pin,
+    dc=dc_pin,
+    rst=reset_pin,
+    baudrate=BAUDRATE,
+    width=135,
+    height=240,
+    x_offset=53,
+    y_offset=40,
+)
+
+
+# these setup the code for our buttons and the backlight and tell the pi to treat the GPIO pins as digitalIO vs analogIO
+backlight = digitalio.DigitalInOut(board.D22)
+backlight.switch_to_output()
+backlight.value = True
+buttonA = digitalio.DigitalInOut(board.D23)
+buttonB = digitalio.DigitalInOut(board.D24)
+buttonA.switch_to_input()
+buttonB.switch_to_input()
+
+
 RESPEAKER_RATE = 16000
 RESPEAKER_CHANNELS = 6 # change base on firmwares, 1_channel_firmware.bin as 1 or 6_channels_firmware.bin as 6
 RESPEAKER_WIDTH = 2
@@ -12,30 +51,39 @@ WAVE_OUTPUT_FILENAME = "output.wav"
 
 p = pyaudio.PyAudio()
 
-stream = p.open(
+# Main loop:
+while True:
+    if buttonB.value and not buttonA.value:  # just button A pressed
+        display.fill(color565(*list(webcolors.name_to_rgb('red')))) # set the screen to the users color
+        stream = p.open(
             rate=RESPEAKER_RATE,
             format=p.get_format_from_width(RESPEAKER_WIDTH),
             channels=RESPEAKER_CHANNELS,
             input=True,
             input_device_index=RESPEAKER_INDEX,)
 
-print("* recording")
+        print("* recording")
 
-frames = []
+        frames = []
 
-for i in range(0, int(RESPEAKER_RATE / CHUNK * RECORD_SECONDS)):
-    data = stream.read(CHUNK)
-    frames.append(data)
+        for i in range(0, int(RESPEAKER_RATE / CHUNK * RECORD_SECONDS)):
+            data = stream.read(CHUNK)
+            frames.append(data)
 
-print("* done recording")
+        print("* done recording")
 
-stream.stop_stream()
-stream.close()
-p.terminate()
+        stream.stop_stream()
+        stream.close()
+        p.terminate()
 
-wf = wave.open(WAVE_OUTPUT_FILENAME, 'wb')
-wf.setnchannels(RESPEAKER_CHANNELS)
-wf.setsampwidth(p.get_sample_size(p.get_format_from_width(RESPEAKER_WIDTH)))
-wf.setframerate(RESPEAKER_RATE)
-wf.writeframes(b''.join(frames))
-wf.close()
+        wf = wave.open(WAVE_OUTPUT_FILENAME, 'wb')
+        wf.setnchannels(RESPEAKER_CHANNELS)
+        wf.setsampwidth(p.get_sample_size(p.get_format_from_width(RESPEAKER_WIDTH)))
+        wf.setframerate(RESPEAKER_RATE)
+        wf.writeframes(b''.join(frames))
+        wf.close()
+
+    if buttonA.value and not buttonB.value:  # just button B pressed
+        display.fill(color565(255, 255, 255))  # set the screen to white
+    if not buttonA.value and not buttonB.value:  # none pressed
+        display.fill(color565(0, 255, 0))  # green
